fixes for #11 removed numpy and opencv from dependencies added --disk flag to use_disk removed audio
