+++ aiortc/mediastreams.py
@@ -1,7 +1,5 @@
 import asyncio
-import cv2
 import math
-import numpy
 
 
 class AudioFrame:
@@ -26,29 +24,6 @@ class VideoFrame:
             self.data = b'\x00' * math.ceil(height * 12 / 8 * width)
         else:
             self.data = data
-
-    @classmethod
-    def from_yuv(cls, height, width, data_yuv):
-        data = data_yuv.tobytes()
-        return cls(width, height, data)
-
-    @classmethod
-    def from_bgr(cls, height, width, data_bgr):
-        data_yuv = cv2.cvtColor(data_bgr, cv2.COLOR_BGR2YUV_YV12)
-        return cls.from_yuv(width, height, data_yuv)
-
-    def to_yuv(self):
-        # truncating the data as a workaround for #10
-        data_len = math.ceil(self.height * 12 / 8 * self.width)
-        data = self.data[0:data_len]
-        data_flat = numpy.frombuffer(data, numpy.uint8)
-        data_yuv = data_flat.reshape((math.ceil(self.height * 12 / 8), self.width))
-        return data_yuv
-
-    def to_bgr(self):
-        data_yuv = self.to_yuv()
-        data_bgr = cv2.cvtColor(data_yuv, cv2.COLOR_YUV2BGR_YV12)
-        return data_bgr
 
 
 class MediaStreamTrack:
